Cleaned up code and moved the custom block definitions and code generators to a js file called custom-blockly-blocks.js. Also fixed the input parameter for the 'ask and wait' block so it matches the expected output.
